% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorOverload.R
\name{$<-.largeList}
\alias{"[[<-.largeList"}
\alias{$<-.largeList}
\title{Overload of operator $<-.}
\usage{
\method{$}{largeList}(x, index) <- value
}
\description{
Overload of operator $<-.
}
\details{
x$A <- 1 is equivalent to x[["A"]] <- 1
}

