% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getList.R
\name{getList}
\alias{getList}
\title{Create a R object and bind with file.}
\usage{
getList(file, verbose = FALSE, truncate = FALSE)
}
\arguments{
\item{file}{Name of file}

\item{verbose}{\code{TRUE/FALSE} Extra info}

\item{truncate}{\code{TRUE/FALSE} Truncate the file or not.}
}
\value{
a R object of class "largeList"
}
\description{
Create a R object and bind with file.
}
\details{
Create a R object of class "largeList" and bind it with a file.
If the file exists, the R object will get the length of list and the names of objects.
If the file does not exist or \code{new == TRUE}, an empty list will be written into 
the given file and then bind it.
}
\examples{
largeListObject <- getList("example.llo", verbose = TRUE, truncate = TRUE)
}
\seealso{
\code{\link{largeList}}
}

