% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\name{getListLength}
\alias{getListLength}
\title{Get number of elements in a list file.}
\usage{
getListLength(file)
}
\arguments{
\item{file}{Name of file.}
}
\value{
An integer.
}
\description{
Get number of elements in a list file.
}
\examples{
list_1 <- list("A" = c(1,2), "B" = "abc", list(1, 2, 3))
saveList(object = list_1, file = "example.llo")
getListLength(file = "example.llo")

}
\seealso{
\code{\link{largeList}}
}
