% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorOverload.R
\name{length.largeList}
\alias{length.largeList}
\title{Overload of function length.}
\usage{
\method{length}{largeList}(x)
}
\arguments{
\item{x}{A largeList object created by \code{\link{getList}}.}
}
\value{
An integer.
}
\description{
Get the length of list stored in file.
}
\seealso{
\code{\link{largeList}}
}

