% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\name{modifyNameInList}
\alias{modifyNameInList}
\title{Modify names of elements in a list file.}
\usage{
modifyNameInList(file, index, name)
}
\arguments{
\item{file}{Name of file.}

\item{index}{A numeric or logical vector.}

\item{name}{A character vector consisting replacement names.}
}
\value{
invisible \code{TRUE} if no error occurs.
}
\description{
Modify names of elements in a list file.
}
\details{
Modify names of elements with given indices by replacement values provided in parameter name. If the length 
of replacement values is shorter than the length of indices, values will be used circularly.
}
\examples{
list_1 <- list("A" = c(1,2), "B" = "abc", list(1, 2, 3))
saveList(object = list_1, file = "example.llo")

# by numeric indices
modifyNameInList(file = "example.llo", index = c(1,2), name = c("AA","BB"))

# by logical indices
modifyNameInList(file = "example.llo", index = c(TRUE, TRUE, FALSE), name = c("AA","BB"))

}
\seealso{
\code{\link{largeList}}
}
