% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorOverload.R
\name{names<-.largeList}
\alias{names<-.largeList}
\title{Overload of operator names<-.}
\usage{
\method{names}{largeList}(x) <- value
}
\arguments{
\item{x}{A largeList object created by \code{\link{getList}}.}

\item{value}{A character vector.}
}
\description{
Overload of operator names<-.
}
\details{
It behaviours the same as a normal list object.
}
\examples{
largelist_object <- getList("example.llo", truncate = TRUE)
largelist_object[[]] <- list("A" = 1, "B" = 2, "C" = 3)  ## assign list to the list file
names(largelist_object) <- c("AA", "BB", "CC")
}
\seealso{
\code{\link{largeList}}
}

