% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorOverload.R
\name{names.largeList}
\alias{names.largeList}
\title{Overload of function names.}
\usage{
\method{names}{largeList}(x)
}
\arguments{
\item{x}{A largeList object created by \code{\link{getList}}.}
}
\value{
A character vector.
}
\description{
Get the names of elements stored in list file.
}
\seealso{
\code{\link{largeList}}
}

