% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorOverload.R
\name{[.largeList}
\alias{[.largeList}
\title{Overload of operator [].}
\usage{
\method{[}{largeList}(largeListObject, index = NULL)
}
\arguments{
\item{largeListObject}{A largeList object created by \code{\link{getList}}.}

\item{index}{A numeric vector or a character vecter.}
}
\value{
a list
}
\description{
Overload of operator [].
}
\details{
It behaviours the same as a normal list object.
}
\seealso{
\code{\link{largeList}}
}

