% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorOverload.R
\name{[[.largeList}
\alias{[[.largeList}
\title{Overload of operator [[]].}
\usage{
\method{[[}{largeList}(x, index = NULL)
}
\arguments{
\item{x}{A largeList object created by \code{\link{getList}}.}

\item{index}{A numeric or character vector of length 1.}
}
\value{
A R object.
}
\description{
Overload of operator [[]].
}
\details{
It behaviours the same as a normal list object.
}
\examples{
largelist_object <- getList("example.llo", truncate = TRUE)
largelist_object[[]] <- list("A" = 1, "B" = 2, "C" = 3)  ## assign list to the list file
largelist_object[[1]] ## get 1
largelist_object[["B"]] ## get 2
}
\seealso{
\code{\link{largeList}}
}

