% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorOverload.R
\name{[[<-.largeList}
\alias{[[<-.largeList}
\title{Overload of operator [[]]<-.}
\usage{
\method{[[}{largeList}(largeListObject, index = NULL) <- value
}
\arguments{
\item{largeListObject}{A largeList object created by \code{\link{getList}}.}

\item{index}{\code{NULL}, a numeric vector or a character vecter.}

\item{value}{\code{NULL}, a vector or a list.}
}
\value{
a largeList object
}
\description{
Overload of operator [[]]<-.
}
\details{
It behaviours the same as a normal list object. If index is not provided, the list file
binding with the largeList object will be truncated and elements in
value wiil be saved to the list file. If value is \code{NULL}, elements with given index 
willbe removed.
}
\seealso{
\code{\link{largeList}}
}

