% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorOverload.R
\name{$<-.largeList}
\alias{$<-.largeList}
\title{Overload of operator $<-.}
\usage{
\method{$}{largeList}(x, index) <- value
}
\arguments{
\item{x}{A largeList object created by \code{\link{getList}}.}

\item{index}{\code{NULL}, a numeric or character vector with lenght 1.}

\item{value}{\code{NULL}, a vector or a list.}
}
\description{
Overload of operator $<-.
}
\details{
x$A <- 1 is equivalent to x[["A"]] <- 1
}
\seealso{
\code{\link{[[<-.largeList}} \code{\link{largeList}}
}

