% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\name{getListName}
\alias{getListName}
\title{Get names of elements in a list file.}
\usage{
getListName(file)
}
\arguments{
\item{file}{Name of file.}
}
\value{
A charater vector.
}
\description{
Get names of elements in a list file.
}
\examples{
list1 <- list("A" = c(1,2), "B" = "abc", list(1, 2, 3))
saveList(list1, "example.llo")
getListName("example.llo")
}
\seealso{
\code{\link{largeList}}
}

