% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorOverload.R
\name{length.largeList}
\alias{length.largeList}
\title{Overload of function length.}
\usage{
\method{length}{largeList}(x)
}
\arguments{
\item{x}{A largeList object created by \code{\link{getList}}.}
}
\value{
An integer.
}
\description{
Get the length of list stored in file.
}
\examples{
largelist_object <- getList("example.llo", truncate = TRUE)
largelist_object[[]] <- list("A" = 1, "B" = 2, "C" = 3)  ## assign list to the list file
length(largelist_object) ## get 3
}
\seealso{
\code{\link{largeList}}
}
