% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorOverload.R
\name{names<-.largeList}
\alias{names<-.largeList}
\title{Overload of operator names<-.}
\usage{
\method{names}{largeList}(x) <- value
}
\arguments{
\item{x}{A largeList object created by \code{\link{getList}}.}

\item{value}{A character vector.}
}
\description{
Overload of operator names<-.
}
\details{
It behaviours the same as a normal list object.
}
\seealso{
\code{\link{largeList}}
}

