% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorOverload.R
\name{names.largeList}
\alias{names.largeList}
\title{Overload of function names.}
\usage{
\method{names}{largeList}(x)
}
\arguments{
\item{x}{A largeList object created by \code{\link{getList}}.}
}
\value{
A character vector.
}
\description{
Get the names of elements stored in list file.
}
\examples{
largelist_object <- getList("example.llo", truncate = TRUE)
largelist_object[[]] <- list("A" = 1, "B" = 2, "C" = 3)  ## assign list to the list file
names(largelist_object) ## get c("A", "B", "C")
}
\seealso{
\code{\link{largeList}}
}
