% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{readList}
\alias{readList}
\title{Get elements from a list file.}
\usage{
readList(file, index = NULL)
}
\arguments{
\item{file}{Name of file.}

\item{index}{\code{NULL}, a numeric vector or a character vecter.}
}
\value{
A list object.
}
\description{
Get elements from a list file.
}
\details{
If no index provided, the whole list will be read. Given index could be a numeric (integer) 
vector or a character vector representing the names. If there exist more then one elements 
corresponding to a given name, , the first matched one will be returned. If there are no elements 
with given name, \code{NULL} will be returned. \cr
Files created by \code{\link{saveRDS}} can't be read.
}
\examples{
list1 <- list("A" = c(1,2), "B" = "abc", list(1, 2, 3))
saveList(list1, "example.llo")
readList("example.llo")
readList("example.llo",c(1, 3))
readList("example.llo",c("A", "B"))
}
\seealso{
\code{\link{largeList}}
}

