% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorOverload.R
\name{[.largeList}
\alias{[.largeList}
\title{Overload of operator [].}
\usage{
\method{[}{largeList}(x, index = NULL)
}
\arguments{
\item{x}{A largeList object created by \code{\link{getList}}.}

\item{index}{A numeric, logical or character vector.}
}
\value{
A list
}
\description{
Overload of operator [].
}
\details{
It behaviours the same as a normal list object.
}
\examples{
largelist_object <- getList("example.llo", truncate = TRUE)
largelist_object[[]] <- list("A" = 1, "B" = 2, "C" = 3)  ## assign list to the list file
largelist_object[c(1, 2)] ## get list("A" = 1, "B" = 2)
largelist_object[c(TRUE, FALSE, TRUE)] ## get list("A" = 1, "C" = 3)
largelist_object[c("A", "C")] ## get list("A" = 1, "C" = 3)
}
\seealso{
\code{\link{largeList}}
}

