% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorOverload.R
\name{[[<-.largeList}
\alias{[[<-.largeList}
\title{Overload of operator [[]]<-.}
\usage{
\method{[[}{largeList}(x, index = NULL) <- value
}
\arguments{
\item{x}{A largeList object created by \code{\link{getList}}.}

\item{index}{\code{NULL}, a numeric vector or a character vecter.}

\item{value}{\code{NULL}, a vector or a list.}
}
\description{
Overload of operator [[]]<-.
}
\details{
It behaviours the same as a normal list object. If index is not provided, the list file
binding with the largeList object will be truncated and elements in
value will be saved to the list file. If value is \code{NULL}, element with given index 
will be removed.
}
\seealso{
\code{\link{largeList}}
}

